requirements.txt
requests
google-genai
google-api-python-client
youtube-transcript-api